wersja admin bez edycji tabeli reservation

+++ src/main/java/pl/mirkowski/lukasz/model/Film.java
@@ -25,7 +25,7 @@ public class Film {
 	@Column(columnDefinition="longtext")
 	private String opis;
 	
-	public long getIdfilm() {
+	public int getIdfilm() {
 		return idfilm;
 	}
 	public void setIdfilm(int idfilm) {
